% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap_summary.R
\name{gap_summary}
\alias{gap_summary}
\title{Get mean, median, and count where a column has values,
split by a grouping factor}
\usage{
gap_summary(df, group, column)
}
\arguments{
\item{df}{Table to be analysed}

\item{group}{Column to be grouped}

\item{column}{Column to be aggregated}
}
\value{
tibble Summarising results
}
\description{
Get mean, median, and count where a column has values,
split by a grouping factor
}
\examples{
gap_summary(iris, Species, Sepal.Width)
}
