% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_distance.R
\name{google_distance}
\alias{google_distance}
\title{Get distance data between two points based on all the travel mode options. Works for many origin points.}
\usage{
google_distance(x, dest, arrival_time, key, ...)
}
\arguments{
\item{x}{A vector of origins in address or postcode format}

\item{dest}{A single destinationin address or postocde format}

\item{arrival_time}{A POSIXct datetime that folks need to arrive by}

\item{key}{A google distance API key}

\item{...}{Additional options to pass to `google_distance()`}
}
\value{
Data.frame containing (typically) 4 rows per input element
}
\description{
Get distance data between two points based on all the travel mode options. Works for many origin points.
}
