% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_vs_actuals.R
\name{pred_vs_actuals}
\alias{pred_vs_actuals}
\title{Predicted vs actuals charts}
\usage{
pred_vs_actuals(df)
}
\arguments{
\item{df}{data.frame containing actual, predicted, and gender}
}
\value{
chart
}
\description{
Predicted vs actuals charts
}
\examples{
pred_vs_actuals(data.frame(actual=1, predicted=1, gender="Female"))
}
